Refactor pipelines into package
